requirements.txt
altair==6.0.0
annotated-types==0.7.0
attrs==25.4.0
bitarray==3.8.0
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
cytoolz==1.1.0
et_xmlfile==2.0.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==6.0.0b1
extra-streamlit-components==0.1.81
Flask==3.1.2
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
hexbytes==1.3.1
httplib2==0.31.0
hyperliquid-python-sdk==0.21.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
msgpack==1.1.2
narwhals==2.15.0
numpy==2.3.5
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
parsimonious==0.10.0
pillow==12.1.0
protobuf==6.33.2
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rlp==4.1.0
rpds-py==0.30.0
rsa==4.9.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
streamlit==1.52.2
streamlit-autorefresh==1.0.1
tenacity==9.1.2
toml==0.10.2
toolz==1.1.0
tornado==6.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
websocket-client==1.9.0
Werkzeug==3.1.5
